requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.5
django-crispy-forms==2.1
django-extensions==3.2.3
django-role-permissions==3.2.0
sqlparse==0.5.0